fix: 외부 의존성은 unit/component test의 커버리지에서 제외한다
