+++ package.json
@@ -26,6 +26,7 @@
     "react": "^18.3.1",
     "react-compiler-runtime": "19.0.0-beta-6fc168f-20241025",
     "react-dom": "^18.3.1",
+    "react-hook-form": "^7.53.1",
     "typescript": "^5.6.3",
     "valibot": "^0.42.1"
   },
